Split proportion helpers into separate modules
